1. 적 유닛의 재출현 및 출현 위치 조정. 2. 유저의 발사속도 및 적군의 무기 발사속도 조정.

+++ src/com/example/game/system/Reload.java
@@ -10,7 +10,7 @@ import com.example.game.R;
  */
 public class Reload {
     public boolean isReloading;
-    public int reloadTime;
+    public double reloadTime;
     public Bitmap reloadBitmap;
 
     public Reload(Context context) {
